feat: Add version update checking info during startup
- Introduce a new `check_update()` function in entry point to check for backend updates
- Create a new `utils.py` module in system API to centralize version-related utility functions
- Move version retrieval and PyPI version checking functions to the new utils module
- Add version comparison using `packaging.version` to detect available updates
- Log update availability with a warning message in the application startup process
